backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-research-assistant-backend"
version = "1.0.0"
description = "FastAPI backend for PDF research paper analysis with RAG/CAG"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "PyPDF2==3.0.1",
    "pdfplumber==0.10.3",
    "chromadb==0.4.24",
    "sentence-transformers==2.2.2",
    "google-generativeai==0.3.0",
    "numpy==1.26.2",
    "requests==2.31.0",
    "aiofiles==23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.12.0",
    "flake8==6.1.0",
]
